requirements.txt
# Pin numpy to a Prophet-compatible version
numpy==1.26.4

# HTTP stack compatible with openai==1.40.0
httpx==0.27.2

fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
python-multipart==0.0.9
boto3==1.34.162
mysql-connector-python==9.0.0
celery==5.4.0
redis==5.0.7
polars==1.5.0
pyarrow==16.1.0
duckdb==1.0.0
pandas==2.2.2
openpyxl==3.1.3
scikit-learn==1.5.1
xgboost==2.1.1
prophet==1.1.5
openai==1.40.0
joblib>=1.3

statsmodels>=0.14
patsy>=0.5

# add AutoGluon TimeSeries (brings in autogluon-core automatically)
autogluon.timeseries>=1.1.0
# optional if you also want tabular classification with AutoGluon:
# autogluon.tabular>=0.8.2